split stops into component
